nt.h: NTSTATUS is 32-bit, so say int rather than long
`typedef long NTSTATUS` is right on the PE targets, where long is 32
bits, and wrong everywhere else. In the native sanitizer build long is
64 bits, so ((NTSTATUS)0xC0000034L) is a *positive* number and
NT_SUCCESS() answers true for every failure status there -- every error
path in the library was unreachable, and the asan build skipped
test/stdlib and test/unistd for exactly that reason.

int is 32-bit on both, and matches Windows' own NTSTATUS, which is a
32-bit LONG. Nothing about the PE ABI changes.

With this and the simulated file system in the previous commit, the
native sanitizer build goes from 6/7 to 10/10: test/stdlib and
test/unistd now run under ASan, UBSan and LSan for the first time, so
the fd table, path translation and the stat/chmod/rename paths are
finally covered by something other than Wine.

Found by the agent implementing fuzz/ntstubs.c, which could not make a
single error path testable until it was fixed.

+++ src/internal/nt.h
@@ -27,7 +27,7 @@
 #define TRUE 1
 #define FALSE 0
 
-typedef long NTSTATUS;
+typedef int NTSTATUS;
 typedef void *HANDLE, **PHANDLE;
 typedef void *PVOID;
 typedef unsigned char BOOLEAN, UCHAR;
